test: setup module dedicated to conformance tests
This is where a part of acceptance tests are supposed to be defined. Tests written here are supposed to be "black box", meaning they know nothing about implementation details and focus instead exclusively on outcomes.
