+++ src/patch-place-break-core/pom.xml
@@ -187,10 +187,5 @@
       <artifactId>testcontainers</artifactId>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>org.threeten</groupId>
-      <artifactId>threeten-extra</artifactId>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 </project>
