+++ src/patch-place-break-cts/src/test/java/fr/djaytan/mc/jrppb/cts/MysqlPatchPlaceBreakApiIntegrationTest.java
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package fr.djaytan.mc.jrppb.api;
+package fr.djaytan.mc.jrppb.cts;
 
 import fr.djaytan.mc.jrppb.commons.test.TestResourcesHelper;
 import java.io.IOException;
@@ -28,34 +28,32 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
-import org.testcontainers.containers.MariaDBContainer;
+import org.testcontainers.containers.MySQLContainer;
 
-class MariadbPatchPlaceBreakApiIntegrationTest extends BasePatchPlaceBreakApiIntegrationTest {
+class MysqlPatchPlaceBreakApiIntegrationTest extends BasePatchPlaceBreakApiIntegrationTest {
 
   private static final int DATABASE_ORIGINAL_PORT = 3306;
   private static final String DATABASE_NAME = "patch_place_break";
 
   @SuppressWarnings("resource") // Reusable containers feature enabled: do not clean-up containers!
-  private static final MariaDBContainer<?> MARIADB_CONTAINER =
-      new MariaDBContainer<>("mariadb:11.1.2-jammy")
-          .withDatabaseName(DATABASE_NAME)
-          .withReuse(true);
+  private static final MySQLContainer<?> MYSQL_CONTAINER =
+      new MySQLContainer<>("mysql:8.1.0-oracle").withDatabaseName(DATABASE_NAME).withReuse(true);
 
   private static final String CONFIG_DATA_SOURCE_FILE_NAME = "dataSource.conf";
 
   @BeforeAll
   static void beforeAll() {
-    MARIADB_CONTAINER.start();
+    MYSQL_CONTAINER.start();
   }
 
   @BeforeEach
   void beforeEach() throws IOException {
-    int dbmsPort = MARIADB_CONTAINER.getMappedPort(DATABASE_ORIGINAL_PORT);
-    String username = MARIADB_CONTAINER.getUsername();
-    String password = MARIADB_CONTAINER.getPassword();
+    int dbmsPort = MYSQL_CONTAINER.getMappedPort(DATABASE_ORIGINAL_PORT);
+    String username = MYSQL_CONTAINER.getUsername();
+    String password = MYSQL_CONTAINER.getPassword();
     String givenDataSourceConfFileContent =
         String.format(
-            TestResourcesHelper.getClassResourceAsString(this.getClass(), "mariadb.conf", false),
+            TestResourcesHelper.getClassResourceAsString(this.getClass(), "mysql.conf", false),
             dbmsPort,
             username,
             password);
