+++ src/patch-place-break-cts/src/test/java/fr/djaytan/mc/jrppb/cts/SqlitePatchPlaceBreakApiTest.java
@@ -20,6 +20,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package fr.djaytan.mc.jrppb.api;
+package fr.djaytan.mc.jrppb.cts;
 
 class SqlitePatchPlaceBreakApiTest extends BasePatchPlaceBreakApiIntegrationTest {}
