+++ src/patch-place-break-cts/src/test/resources/fr/djaytan/mc/jrppb/cts/MysqlPatchPlaceBreakApiIntegrationTest/mysql.conf
